change in register page
