+++ lib/view/main_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:final_project/properties_app/prop.dart';
 import 'package:convex_bottom_bar/convex_bottom_bar.dart';
 import 'package:final_project/view/home.dart';
-import 'package:final_project/view/profile.dart';
+import 'package:final_project/view/profile_page/profile.dart';
 import 'package:final_project/view/transaction.dart';
 import 'package:final_project/view/wishlist.dart';
