+++ lib/view/onboarding_page/component/intro_two.dart
