+++ pubspec.yaml
@@ -31,6 +31,7 @@ environment:
 dependencies:
   convex_bottom_bar: ^3.1.0+1
   cupertino_icons: ^1.0.2
+  email_validator: ^2.1.17
   flutter:
     sdk: flutter
   get: ^4.6.5
